GenAI_Travel_Planner_Clean/requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0

# AI and Language Models
openai>=1.0.0
google-genai>=0.2.0

# AI Agent Framework
agno>=0.1.0

# Travel APIs
amadeus>=8.0.0
googlemaps>=4.10.0

# Search APIs
google-search-results>=2.4.0

# Data Processing and HTTP Requests
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# Date and Time Handling
python-dateutil>=2.8.0

# Environment Variables
python-dotenv>=1.0.0

# JSON and Data Handling
json5>=0.9.0

# Google AI dependencies
google-cloud-aiplatform>=1.0.0
google-generativeai>=0.3.0

# HTTP and URL handling
urllib3>=2.0.0

# Additional dependencies that might be needed
typing-extensions>=4.0.0
protobuf>=3.19.0
